update dashboard modal, migration, controller, serivce

+++ src/controlles/dashboard.controller.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { dashboardService } from '../services/dashboard.service.js';
-import { Controller } from '../types.js';
+import { Controller, UpdatedFields } from '../types.js';
 
 class DashboardController {
   private static instance: DashboardController | null = null;
@@ -21,7 +21,7 @@ class DashboardController {
     const { title } = req.body;
 
     if (!title) {
-      res.status(400).json({ message: 'Title is requeried' });
+      res.status(400).json({ message: 'Title is required' });
 
       return;
     }
@@ -31,8 +31,26 @@ class DashboardController {
     res.status(200).json(dashboard);
   };
 
-  getAllDashboards: Controller = async (req, res) => {
-    const dashboardData = await dashboardService.getAllDashboards();
+  getDashboards: Controller = async (req, res) => {
+    const { isOpened } = req.query;
+
+    if (!isOpened) {
+      res.status(400).json({ message: 'The isOpened parameter is required' });
+
+      return;
+    }
+
+    if (isOpened !== 'true' && isOpened !== 'false') {
+      res.status(400).json({
+        message: 'The isOpened parameter can have only one parameter true or false',
+      });
+
+      return;
+    }
+
+    const isOpenedBoolean = isOpened === 'true';
+
+    const dashboardData = await dashboardService.getDashboards(isOpenedBoolean);
 
     res.status(200).json(dashboardData);
   };
@@ -48,24 +66,86 @@ class DashboardController {
   };
 
   updateDashboard: Controller = async (req, res) => {
-    const { title } = req.body;
+    const { title, isOpened, important } = req.body;
     const { id } = req.params;
+
+    if (!id) {
+      res.status(400).json({ message: 'id parameter is required' });
+
+      return;
+    }
+
+    if (!title && !isOpened && !important) {
+      res.status(400).json({
+        message: 'At least one field (title, isOpened, important) must be provided for updating',
+      });
+
+      return;
+    }
+
+    if (isOpened && isOpened !== 'true' && isOpened !== 'false') {
+      res.status(400).json({
+        message: 'Field isOpened can has only one of two value true or false',
+      });
+
+      return;
+    }
+
+    if (important && important !== 'true' && important !== 'false') {
+      res.status(400).json({
+        message: 'Field important can has only one of two value true or false',
+      });
+    }
+
+    const updatedFields: UpdatedFields = {};
+
+    if (title) {
+      updatedFields.title = title;
+    }
+
+    if (isOpened) {
+      updatedFields.isOpened = isOpened === 'true';
+    }
+
+    if (important) {
+      updatedFields.important = important === 'true';
+    }
 
     const normalizeId = Number(id);
 
-    const dashboardData = await dashboardService.updateDashboard(normalizeId, title);
+    const dashboardData = await dashboardService.updateDashboard(normalizeId, updatedFields);
 
-    res.status(200).json(dashboardData);
+    if (dashboardData[0] === 0) {
+      res.status(404).json({ message: `Dashboard with id ${normalizeId} not found` });
+
+      return;
+    }
+
+    res.status(200).json({ message: 'Dashboard updated successfully' });
   };
 
   deleteDashboard: Controller = async (req, res) => {
     const { id } = req.params;
 
+    if (!id) {
+      res.status(400).json({ message: 'id parameter is required' });
+
+      return;
+    }
+
     const normalizeId = Number(id);
 
     const dashboardData = await dashboardService.deleteDashboard(normalizeId);
 
-    res.status(200).json(dashboardData);
+    if (dashboardData === 0) {
+      res.status(404).json({ message: `Dashboard with id ${normalizeId} not found` });
+
+      return;
+    }
+
+    console.log(dashboardData);
+
+    res.status(200).json({ message: 'Dashboard deleted successfully' });
   };
 }
